WIP: refactor diceroller
