+++ src/context/DDDiceContext.tsx
@@ -1,17 +1,10 @@
-import { createContext, useContext } from "react";
 import { ThreeDDice } from "dddice-js";
+import { create } from "zustand";
 
-export type DDDiceContextType = {
-    dice: ThreeDDice | undefined;
+export type DiceRoller = {
+    roller: ThreeDDice;
 };
 
-export const DiceContext = createContext<DDDiceContextType | undefined>(undefined);
-
-export const useDiceContext = (): DDDiceContextType => {
-    const diceContext = useContext(DiceContext);
-    if (diceContext === undefined) {
-        throw new Error("DDDice not ready");
-    }
-
-    return diceContext;
-};
+export const useDiceRoller = create<DiceRoller>()(() => ({
+    roller: new ThreeDDice(),
+}));
