+++ src/css/components/_dddice.scss
@@ -1,13 +1,8 @@
 #DiceCanvas {
-    display: none;
     position: absolute;
     top: 0;
     left: 0;
     z-index: -1;
     width: 100svw;
     height: 100svh;
-
-    &.visible {
-        display: block;
-    }
 }
